@@ -1,29 +1,30 @@
 class TeamsController < ApplicationController
-  before_filter :authenticate_user!, only: [:create, :update, :destroy]
+  before_action :authenticate_user!, only: [:create, :update, :destroy]
 
   def index
-    respond_with Team.all
+    render json: Team.all
   end
-  
+
   def create
-    respond_with Team.create(team_params)
+    render json: Team.create(team_params)
   end
-  
+
   def show
-    respond_with Team.find(params[:id])
+    render json: Team.find(params[:id])
   end
-  
+
   def update
     team = Team.find(params[:id])
-    respond_with team.update_attributes(team_params)
+    render json: team.update_attributes(team_params)
   end
-  
+
   def destroy
-    respond_with Team.delete(params[:id])
+    render json: Team.delete(params[:id])
   end
-  
+
   private
-    def team_params
-      params.require(:team).permit(:name, :nhlID)
-    end
+
+  def team_params
+    params.require(:team).permit(:name, :nhlID)
+  end
 end
