@@ -1,6 +1,6 @@
-class Team < ActiveRecord::Base
-  has_many :players, -> { order "id ASC" }
-  
+class Team < ApplicationRecord
+  has_many :players, -> { order 'id ASC' }
+
   def as_json(options = {})
     super(options.merge(include: :players))
   end
