@@ -1,11 +1,3 @@
-class ApplicationController < ActionController::Base
-  # Prevent CSRF attacks by raising an exception.
-  # For APIs, you may want to use :null_session instead.
-  protect_from_forgery with: :exception
-  
+class ApplicationController < ActionController::API
   respond_to :json
-  
-  def angular
-    render 'layouts/application'
-  end
 end
