@@ -1,17 +1,16 @@
 Rails.application.routes.draw do
-  # The priority is based upon order of creation: first created -> highest priority.
-  # See how all your routes lay out with "rake routes".
-  root 'application#angular'
-  
-  resources :teams, defaults: { format: 'json' }, only: [:index, :create, :show, :update, :destroy] do
+  # For details on the DSL available within this file, see http://guides.rubyonrails.org/routing.html
+  root 'static#base'
+
+  resources :teams, only: [:index, :create, :show, :update, :destroy] do
     resources :players, shallow: true, only: [:index, :create, :show, :update, :destroy]
   end
-  
-  resources :entries, defaults: { format: 'json' }, only: [:index, :create, :show, :update, :destroy]
-  
+
+  resources :entries, only: [:index, :create, :show, :update, :destroy]
+
   devise_for :users, defaults: { format: 'json' }
-  
-  match '/update_player_stats', to: 'entries#update_player_stats', via: 'get', defaults: { format: 'json' }
-  
-  get '*path' => 'application#angular'
+
+  match '/update_player_stats', to: 'entries#update_player_stats', via: 'get'
+
+  get '*path' => 'static#base'
 end
